Moving Combobox subclasses to their own files
