+++ src/System.Windows.Forms/src/System.Windows.Forms.csproj
@@ -629,4 +629,49 @@
       <LogicalName>System.Windows.Forms.blank</LogicalName>
     </EmbeddedResource>
   </ItemGroup>
+
+  <ItemGroup>
+    <Compile Update="System\Windows\Forms\ComboBox.ChildAccessibleObject.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ObjectCollection.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.FlatComboAdapter.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.AutoCompleteDropDownFinder.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ACNativeWindow.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ComboBoxChildDropDownButtonUiaProvider.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ComboBoxChildTextUiaProvider.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ComboBoxChildListUiaProvider.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ComboBoxChildEditUiaProvider.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ComboBoxAccessibleObject.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ComboBoxItemAccessibleObjectCollection.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ComboBoxItemAccessibleObject.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ItemComparer.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Update="System\Windows\Forms\ComboBox.ComboBoxChildNativeWindow.cs">
+      <SubType>Component</SubType>
+    </Compile>
+  </ItemGroup>
 </Project>
