+++ src/System.Windows.Forms/src/System/Windows/Forms/ComboBox.cs
@@ -12,11 +12,8 @@ using System.Drawing;
 using System.Drawing.Design;
 using System.Globalization;
 using System.Runtime.InteropServices;
-using System.Runtime.Serialization;
-using System.Text;
 using System.Threading;
 using System.Windows.Forms.Layout;
-using Accessibility;
 using static Interop;
 using static Interop.User32;
 
@@ -31,7 +28,7 @@ namespace System.Windows.Forms
     [DefaultBindingProperty(nameof(Text))]
     [Designer("System.Windows.Forms.Design.ComboBoxDesigner, " + AssemblyRef.SystemDesign)]
     [SRDescription(nameof(SR.DescriptionComboBox))]
-    public class ComboBox : ListControl
+    public partial class ComboBox : ListControl
     {
         private static readonly object EVENT_DROPDOWN = new object();
         private static readonly object EVENT_DRAWITEM = new object();
@@ -3911,2244 +3908,6 @@ namespace System.Windows.Forms
             }
         }
 
-        private class ComboBoxChildNativeWindow : NativeWindow
-        {
-            private readonly ComboBox _owner;
-            private InternalAccessibleObject _accessibilityObject;
-            private readonly ChildWindowType _childWindowType;
-
-            public ComboBoxChildNativeWindow(ComboBox comboBox, ChildWindowType childWindowType)
-            {
-                _owner = comboBox;
-                _childWindowType = childWindowType;
-            }
-
-            protected override void WndProc(ref Message m)
-            {
-                switch ((User32.WM)m.Msg)
-                {
-                    case WM.GETOBJECT:
-                        WmGetObject(ref m);
-                        return;
-                    case WM.MOUSEMOVE:
-                        if (_childWindowType == ChildWindowType.DropDownList)
-                        {
-                            // Need to track the selection change via mouse over to
-                            // raise focus changed event for the items. Monitoring
-                            // item change in setters does not guarantee that focus
-                            // is properly announced.
-                            object before = _owner.SelectedItem;
-                            DefWndProc(ref m);
-                            object after = _owner.SelectedItem;
-                            if (before != after)
-                            {
-                                (_owner.AccessibilityObject as ComboBoxAccessibleObject).SetComboBoxItemFocus();
-                            }
-                        }
-                        else
-                        {
-                            _owner.ChildWndProc(ref m);
-                        }
-                        break;
-                    default:
-                        if (_childWindowType == ChildWindowType.DropDownList)
-                        {
-                            DefWndProc(ref m); // Drop Down window should behave by its own.
-                        }
-                        else
-                        {
-                            _owner.ChildWndProc(ref m);
-                        }
-                        break;
-                }
-            }
-
-            private ChildAccessibleObject GetChildAccessibleObject(ChildWindowType childWindowType)
-            {
-                if (childWindowType == ChildWindowType.Edit)
-                {
-                    return _owner.ChildEditAccessibleObject;
-                }
-                else if (childWindowType == ChildWindowType.ListBox || childWindowType == ChildWindowType.DropDownList)
-                {
-                    return _owner.ChildListAccessibleObject;
-                }
-
-                return new ChildAccessibleObject(_owner, Handle);
-            }
-
-            private void WmGetObject(ref Message m)
-            {
-                if (m.LParam == (IntPtr)NativeMethods.UiaRootObjectId &&
-                    // Managed UIAutomation providers are supplied for child list windows but not for the child edit window.
-                    // Child list accessibility object provides all necessary patterns and UIAutomation notifications,
-                    // so there is no need to native provider supplement.
-                    // Child edit accessibility object has only partial support of edit box accessibility, most of the patterns
-                    // and notifications for child edit window are supplied by native providers, so here is no need to
-                    // override root UIA object for child edit window.
-                    (_childWindowType == ChildWindowType.ListBox || _childWindowType == ChildWindowType.DropDownList))
-                {
-                    AccessibleObject uiaProvider = GetChildAccessibleObject(_childWindowType);
-
-                    // If the requested object identifier is UiaRootObjectId,
-                    // we should return an UI Automation provider using the UiaReturnRawElementProvider function.
-                    InternalAccessibleObject internalAccessibleObject = new InternalAccessibleObject(uiaProvider);
-                    m.Result = UiaCore.UiaReturnRawElementProvider(
-                        new HandleRef(this, Handle),
-                        m.WParam,
-                        m.LParam,
-                        internalAccessibleObject);
-
-                    return;
-                }
-
-                // See "How to Handle WM_GETOBJECT" in MSDN
-                //
-                if (unchecked((int)(long)m.LParam) == OBJID.CLIENT)
-                {
-                    // Get the IAccessible GUID
-                    //
-                    Guid IID_IAccessible = new Guid(NativeMethods.uuid_IAccessible);
-
-                    // Get an Lresult for the accessibility Object for this control
-                    //
-                    IntPtr punkAcc;
-                    try
-                    {
-                        AccessibleObject wfAccessibleObject = null;
-                        UiaCore.IAccessibleInternal iacc = null;
-
-                        if (_accessibilityObject is null)
-                        {
-                            wfAccessibleObject = GetChildAccessibleObject(_childWindowType);
-                            _accessibilityObject = new InternalAccessibleObject(wfAccessibleObject);
-                        }
-                        iacc = (UiaCore.IAccessibleInternal)_accessibilityObject;
-
-                        // Obtain the Lresult
-                        //
-                        punkAcc = Marshal.GetIUnknownForObject(iacc);
-
-                        try
-                        {
-                            m.Result = Oleacc.LresultFromObject(ref IID_IAccessible, m.WParam, new HandleRef(this, punkAcc));
-                        }
-                        finally
-                        {
-                            Marshal.Release(punkAcc);
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        throw new InvalidOperationException(SR.RichControlLresult, e);
-                    }
-                }
-                else
-                {  // m.lparam != OBJID_CLIENT, so do default message processing
-                    DefWndProc(ref m);
-                }
-            }
-        }
-
-        private sealed class ItemComparer : IComparer
-        {
-            private readonly ComboBox _comboBox;
-
-            public ItemComparer(ComboBox comboBox)
-            {
-                _comboBox = comboBox;
-            }
-
-            public int Compare(object item1, object item2)
-            {
-                if (item1 is null)
-                {
-                    if (item2 is null)
-                    {
-                        return 0; //both null, then they are equal
-                    }
-
-                    return -1; //item1 is null, but item2 is valid (greater)
-                }
-                if (item2 is null)
-                {
-                    return 1; //item2 is null, so item 1 is greater
-                }
-
-                string itemName1 = _comboBox.GetItemText(item1);
-                string itemName2 = _comboBox.GetItemText(item2);
-
-                CompareInfo compInfo = (Application.CurrentCulture).CompareInfo;
-                return compInfo.Compare(itemName1, itemName2, CompareOptions.StringSort);
-            }
-        }
-
-        [ListBindable(false)]
-        public class ObjectCollection : IList
-        {
-            private readonly ComboBox _owner;
-            private ArrayList _innerList;
-            private IComparer _comparer;
-
-            public ObjectCollection(ComboBox owner)
-            {
-                _owner = owner;
-            }
-
-            private IComparer Comparer
-            {
-                get
-                {
-                    if (_comparer is null)
-                    {
-                        _comparer = new ItemComparer(_owner);
-                    }
-                    return _comparer;
-                }
-            }
-
-            private ArrayList InnerList
-            {
-                get
-                {
-                    if (_innerList is null)
-                    {
-                        _innerList = new ArrayList();
-                    }
-                    return _innerList;
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the number of items.
-            /// </summary>
-            public int Count
-            {
-                get
-                {
-                    return InnerList.Count;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Adds an item to the combo box. For an unsorted combo box, the item is
-            ///  added to the end of the existing list of items. For a sorted combo box,
-            ///  the item is inserted into the list according to its sorted position.
-            ///  The item's toString() method is called to obtain the string that is
-            ///  displayed in the combo box.
-            ///  A SystemException occurs if there is insufficient space available to
-            ///  store the new item.
-            /// </summary>
-            public int Add(object item)
-            {
-                _owner.CheckNoDataSource();
-                int index = AddInternal(item);
-                if (_owner.UpdateNeeded() && _owner.AutoCompleteSource == AutoCompleteSource.ListItems)
-                {
-                    _owner.SetAutoComplete(false, false);
-                }
-                return index;
-            }
-
-            private int AddInternal(object item)
-            {
-                if (item is null)
-                {
-                    throw new ArgumentNullException(nameof(item));
-                }
-                int index = -1;
-                if (!_owner._sorted)
-                {
-                    InnerList.Add(item);
-                }
-                else
-                {
-                    index = InnerList.BinarySearch(item, Comparer);
-                    if (index < 0)
-                    {
-                        index = ~index; // getting the index of the first element that is larger than the search value
-                    }
-
-                    Debug.Assert(index >= 0 && index <= InnerList.Count, "Wrong index for insert");
-                    InnerList.Insert(index, item);
-                }
-                bool successful = false;
-
-                try
-                {
-                    if (_owner._sorted)
-                    {
-                        if (_owner.IsHandleCreated)
-                        {
-                            _owner.NativeInsert(index, item);
-                        }
-                    }
-                    else
-                    {
-                        index = InnerList.Count - 1;
-                        if (_owner.IsHandleCreated)
-                        {
-                            _owner.NativeAdd(item);
-                        }
-                    }
-                    successful = true;
-                }
-                finally
-                {
-                    if (!successful)
-                    {
-                        InnerList.Remove(item);
-                    }
-                }
-
-                return index;
-            }
-
-            int IList.Add(object item)
-            {
-                return Add(item);
-            }
-
-            public void AddRange(object[] items)
-            {
-                _owner.CheckNoDataSource();
-                _owner.BeginUpdate();
-                try
-                {
-                    AddRangeInternal(items);
-                }
-                finally
-                {
-                    _owner.EndUpdate();
-                }
-            }
-
-            internal void AddRangeInternal(IList items)
-            {
-                if (items is null)
-                {
-                    throw new ArgumentNullException(nameof(items));
-                }
-                foreach (object item in items)
-                {
-                    // adding items one-by-one for performance (especially for sorted combobox)
-                    // we can not rely on ArrayList.Sort since its worst case complexity is n*n
-                    // AddInternal is based on BinarySearch and ensures n*log(n) complexity
-                    AddInternal(item);
-                }
-                if (_owner.AutoCompleteSource == AutoCompleteSource.ListItems)
-                {
-                    _owner.SetAutoComplete(false, false);
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the item with the specified index.
-            /// </summary>
-            [Browsable(false)]
-            [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
-            public virtual object this[int index]
-            {
-                get
-                {
-                    if (index < 0 || index >= InnerList.Count)
-                    {
-                        throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                    }
-
-                    return InnerList[index];
-                }
-                set
-                {
-                    _owner.CheckNoDataSource();
-                    SetItemInternal(index, value);
-                }
-            }
-
-            /// <summary>
-            ///  Removes all items from the ComboBox.
-            /// </summary>
-            public void Clear()
-            {
-                _owner.CheckNoDataSource();
-                ClearInternal();
-            }
-
-            internal void ClearInternal()
-            {
-                if (_owner.IsHandleCreated)
-                {
-                    _owner.NativeClear();
-                }
-
-                InnerList.Clear();
-                _owner._selectedIndex = -1;
-                if (_owner.AutoCompleteSource == AutoCompleteSource.ListItems)
-                {
-                    _owner.SetAutoComplete(false, true /*recreateHandle*/);
-                }
-            }
-
-            public bool Contains(object value)
-            {
-                return IndexOf(value) != -1;
-            }
-
-            /// <summary>
-            ///  Copies the ComboBox Items collection to a destination array.
-            /// </summary>
-            public void CopyTo(object[] destination, int arrayIndex)
-            {
-                InnerList.CopyTo(destination, arrayIndex);
-            }
-
-            void ICollection.CopyTo(Array destination, int index)
-            {
-                InnerList.CopyTo(destination, index);
-            }
-
-            /// <summary>
-            ///  Returns an enumerator for the ComboBox Items collection.
-            /// </summary>
-            public IEnumerator GetEnumerator()
-            {
-                return InnerList.GetEnumerator();
-            }
-
-            public int IndexOf(object value)
-            {
-                if (value is null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                return InnerList.IndexOf(value);
-            }
-
-            /// <summary>
-            ///  Adds an item to the combo box. For an unsorted combo box, the item is
-            ///  added to the end of the existing list of items. For a sorted combo box,
-            ///  the item is inserted into the list according to its sorted position.
-            ///  The item's toString() method is called to obtain the string that is
-            ///  displayed in the combo box.
-            ///  A SystemException occurs if there is insufficient space available to
-            ///  store the new item.
-            /// </summary>
-            public void Insert(int index, object item)
-            {
-                _owner.CheckNoDataSource();
-
-                if (item is null)
-                {
-                    throw new ArgumentNullException(nameof(item));
-                }
-
-                if (index < 0 || index > InnerList.Count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                // If the combo box is sorted, then nust treat this like an add
-                // because we are going to twiddle the index anyway.
-                //
-                if (_owner._sorted)
-                {
-                    Add(item);
-                }
-                else
-                {
-                    InnerList.Insert(index, item);
-                    if (_owner.IsHandleCreated)
-                    {
-                        bool successful = false;
-
-                        try
-                        {
-                            _owner.NativeInsert(index, item);
-                            successful = true;
-                        }
-                        finally
-                        {
-                            if (successful)
-                            {
-                                if (_owner.AutoCompleteSource == AutoCompleteSource.ListItems)
-                                {
-                                    _owner.SetAutoComplete(false, false);
-                                }
-                            }
-                            else
-                            {
-                                InnerList.RemoveAt(index);
-                            }
-                        }
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Removes an item from the ComboBox at the given index.
-            /// </summary>
-            public void RemoveAt(int index)
-            {
-                _owner.CheckNoDataSource();
-
-                if (index < 0 || index >= InnerList.Count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                if (_owner.IsHandleCreated)
-                {
-                    _owner.NativeRemoveAt(index);
-                }
-
-                InnerList.RemoveAt(index);
-                if (!_owner.IsHandleCreated && index < _owner._selectedIndex)
-                {
-                    _owner._selectedIndex--;
-                }
-                if (_owner.AutoCompleteSource == AutoCompleteSource.ListItems)
-                {
-                    _owner.SetAutoComplete(false, false);
-                }
-            }
-
-            /// <summary>
-            ///  Removes the given item from the ComboBox, provided that it is
-            ///  actually in the list.
-            /// </summary>
-            public void Remove(object value)
-            {
-                int index = InnerList.IndexOf(value);
-
-                if (index != -1)
-                {
-                    RemoveAt(index);
-                }
-            }
-
-            internal void SetItemInternal(int index, object value)
-            {
-                if (index < 0 || index >= InnerList.Count)
-                {
-                    throw new ArgumentOutOfRangeException(nameof(index), index, string.Format(SR.InvalidArgument, nameof(index), index));
-                }
-
-                InnerList[index] = value ?? throw new ArgumentNullException(nameof(value));
-
-                // If the native control has been created, and the display text of the new list item object
-                // is different to the current text in the native list item, recreate the native list item...
-                if (_owner.IsHandleCreated)
-                {
-                    bool selected = (index == _owner.SelectedIndex);
-
-                    if (string.Compare(_owner.GetItemText(value), _owner.NativeGetItemText(index), true, CultureInfo.CurrentCulture) != 0)
-                    {
-                        _owner.NativeRemoveAt(index);
-                        _owner.NativeInsert(index, value);
-                        if (selected)
-                        {
-                            _owner.SelectedIndex = index;
-                            _owner.UpdateText();
-                        }
-                        if (_owner.AutoCompleteSource == AutoCompleteSource.ListItems)
-                        {
-                            _owner.SetAutoComplete(false, false);
-                        }
-                    }
-                    else
-                    {
-                        // NEW - FOR COMPATIBILITY REASONS
-                        // Minimum compatibility fix
-                        if (selected)
-                        {
-                            _owner.OnSelectedItemChanged(EventArgs.Empty);   //we do this because set_SelectedIndex does this. (for consistency)
-                            _owner.OnSelectedIndexChanged(EventArgs.Empty);
-                        }
-                    }
-                }
-            }
-        } // end ObjectCollection
-
-        public class ChildAccessibleObject : AccessibleObject
-        {
-            private readonly ComboBox _owner;
-
-            public ChildAccessibleObject(ComboBox owner, IntPtr handle)
-            {
-                Debug.Assert(owner?.IsHandleCreated is true, "ComboBox's handle hasn't been created");
-
-                _owner = owner;
-
-                if (owner?.IsHandleCreated is true)
-                {
-                    UseStdAccessibleObjects(handle);
-                }
-            }
-
-            public override string Name
-            {
-                get
-                {
-                    return _owner.AccessibilityObject.Name;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  Represents the ComboBox item accessible object.
-        /// </summary>
-        internal class ComboBoxItemAccessibleObject : AccessibleObject
-        {
-            private readonly ComboBox _owningComboBox;
-            private readonly object _owningItem;
-            private IAccessible _systemIAccessible;
-
-            /// <summary>
-            ///  Initializes new instance of ComboBox item accessible object.
-            /// </summary>
-            /// <param name="owningComboBox">The owning ComboBox.</param>
-            /// <param name="owningItem">The owning ComboBox item.</param>
-            public ComboBoxItemAccessibleObject(ComboBox owningComboBox, object owningItem)
-            {
-                _owningComboBox = owningComboBox;
-                _owningItem = owningItem;
-
-                _systemIAccessible = _owningComboBox.ChildListAccessibleObject.GetSystemIAccessibleInternal();
-            }
-
-            /// <summary>
-            ///  Gets the ComboBox Item bounds.
-            /// </summary>
-            public override Rectangle Bounds
-            {
-                get
-                {
-                    ChildAccessibleObject listAccessibleObject = _owningComboBox.ChildListAccessibleObject;
-                    int currentIndex = GetCurrentIndex();
-
-                    Rectangle parentRect = listAccessibleObject.BoundingRectangle;
-                    int left = parentRect.Left;
-                    int top = parentRect.Top + _owningComboBox.ItemHeight * currentIndex;
-                    int width = parentRect.Width;
-                    int height = _owningComboBox.ItemHeight;
-
-                    return new Rectangle(left, top, width, height);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the ComboBox item default action.
-            /// </summary>
-            public override string DefaultAction => _systemIAccessible.accDefaultAction[GetChildId()];
-
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                switch (direction)
-                {
-                    case UiaCore.NavigateDirection.Parent:
-                        return _owningComboBox.ChildListAccessibleObject;
-                    case UiaCore.NavigateDirection.NextSibling:
-                        int currentIndex = GetCurrentIndex();
-                        if (_owningComboBox.ChildListAccessibleObject is ComboBoxChildListUiaProvider comboBoxChildListUiaProvider)
-                        {
-                            int itemsCount = comboBoxChildListUiaProvider.GetChildFragmentCount();
-                            int nextItemIndex = currentIndex + 1;
-                            if (itemsCount > nextItemIndex)
-                            {
-                                return comboBoxChildListUiaProvider.GetChildFragment(nextItemIndex);
-                            }
-                        }
-                        break;
-                    case UiaCore.NavigateDirection.PreviousSibling:
-                        currentIndex = GetCurrentIndex();
-                        comboBoxChildListUiaProvider = _owningComboBox.ChildListAccessibleObject as ComboBoxChildListUiaProvider;
-                        if (comboBoxChildListUiaProvider != null)
-                        {
-                            var itemsCount = comboBoxChildListUiaProvider.GetChildFragmentCount();
-                            int previousItemIndex = currentIndex - 1;
-                            if (previousItemIndex >= 0)
-                            {
-                                return comboBoxChildListUiaProvider.GetChildFragment(previousItemIndex);
-                            }
-                        }
-
-                        break;
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owningComboBox.AccessibilityObject;
-                }
-            }
-
-            private int GetCurrentIndex()
-            {
-                return _owningComboBox.Items.IndexOf(_owningItem);
-            }
-
-            internal override int GetChildId()
-            {
-                return GetCurrentIndex() + 1; // Index is zero-based, Child ID is 1-based.
-            }
-
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.RuntimeIdPropertyId:
-                        return RuntimeId;
-                    case UiaCore.UIA.BoundingRectanglePropertyId:
-                        return BoundingRectangle;
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.ListItemControlTypeId;
-                    case UiaCore.UIA.NamePropertyId:
-                        return Name;
-                    case UiaCore.UIA.AccessKeyPropertyId:
-                        return KeyboardShortcut ?? string.Empty;
-                    case UiaCore.UIA.HasKeyboardFocusPropertyId:
-                        return _owningComboBox.Focused && _owningComboBox.SelectedIndex == GetCurrentIndex();
-                    case UiaCore.UIA.IsKeyboardFocusablePropertyId:
-                        return (State & AccessibleStates.Focusable) == AccessibleStates.Focusable;
-                    case UiaCore.UIA.IsEnabledPropertyId:
-                        return _owningComboBox.Enabled;
-                    case UiaCore.UIA.HelpTextPropertyId:
-                        return Help ?? string.Empty;
-                    case UiaCore.UIA.IsControlElementPropertyId:
-                        return true;
-                    case UiaCore.UIA.IsContentElementPropertyId:
-                        return true;
-                    case UiaCore.UIA.IsPasswordPropertyId:
-                        return false;
-                    case UiaCore.UIA.IsOffscreenPropertyId:
-                        return (State & AccessibleStates.Offscreen) == AccessibleStates.Offscreen;
-                    case UiaCore.UIA.IsSelectionItemPatternAvailablePropertyId:
-                        return true;
-                    case UiaCore.UIA.SelectionItemIsSelectedPropertyId:
-                        return (State & AccessibleStates.Selected) != 0;
-                    case UiaCore.UIA.SelectionItemSelectionContainerPropertyId:
-                        return _owningComboBox.ChildListAccessibleObject;
-
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the help text.
-            /// </summary>
-            public override string Help => _systemIAccessible.accHelp[GetChildId()];
-
-            /// <summary>
-            ///  Indicates whether specified pattern is supported.
-            /// </summary>
-            /// <param name="patternId">The pattern ID.</param>
-            /// <returns>True if specified </returns>
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.LegacyIAccessiblePatternId ||
-                    patternId == UiaCore.UIA.InvokePatternId ||
-                    patternId == UiaCore.UIA.SelectionItemPatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            /// <summary>
-            ///  Gets or sets the accessible name.
-            /// </summary>
-            public override string Name
-            {
-                get
-                {
-                    if (_owningComboBox != null)
-                    {
-                        return _owningComboBox.GetItemText(_owningItem);
-                    }
-
-                    return base.Name;
-                }
-                set => base.Name = value;
-            }
-
-            /// <summary>
-            ///  Gets the accessible role.
-            /// </summary>
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    var accRole = _systemIAccessible.get_accRole(GetChildId());
-                    return accRole != null
-                        ? (AccessibleRole)accRole
-                        : AccessibleRole.None;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the runtime ID.
-            /// </summary>
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    if (!_owningComboBox.IsHandleCreated)
-                    {
-                        return base.RuntimeId;
-                    }
-
-                    var runtimeId = new int[4];
-                    runtimeId[0] = RuntimeIDFirstItem;
-                    runtimeId[1] = (int)(long)_owningComboBox.InternalHandle;
-                    runtimeId[2] = _owningComboBox.GetListNativeWindowRuntimeIdPart();
-
-                    var comboBoxAccessibleObject = _owningComboBox.AccessibilityObject as ComboBoxAccessibleObject;
-                    runtimeId[3] = comboBoxAccessibleObject.ItemAccessibleObjects.GetId(_owningItem);
-
-                    return runtimeId;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible state.
-            /// </summary>
-            public override AccessibleStates State
-            {
-                get
-                {
-                    var accState = _systemIAccessible.get_accState(GetChildId());
-                    return accState != null
-                        ? (AccessibleStates)accState
-                        : AccessibleStates.None;
-                }
-            }
-
-            internal override void SetFocus()
-            {
-                RaiseAutomationEvent(UiaCore.UIA.AutomationFocusChangedEventId);
-
-                base.SetFocus();
-            }
-
-            internal unsafe override void SelectItem()
-            {
-                if (!_owningComboBox.IsHandleCreated)
-                {
-                    return;
-                }
-
-                _owningComboBox.SelectedIndex = GetCurrentIndex();
-                InvalidateRect(new HandleRef(this, _owningComboBox.GetListHandle()), null, BOOL.FALSE);
-            }
-
-            internal override void AddToSelection()
-            {
-                if (!_owningComboBox.IsHandleCreated)
-                {
-                    return;
-                }
-
-                SelectItem();
-            }
-
-            internal override void RemoveFromSelection()
-            {
-                // Do nothing, C++ implementation returns UIA_E_INVALIDOPERATION 0x80131509
-            }
-
-            internal override bool IsItemSelected
-            {
-                get
-                {
-                    return (State & AccessibleStates.Selected) != 0;
-                }
-            }
-
-            internal override UiaCore.IRawElementProviderSimple ItemSelectionContainer
-            {
-                get
-                {
-                    return _owningComboBox.ChildListAccessibleObject;
-                }
-            }
-        }
-
-        internal class ComboBoxItemAccessibleObjectCollection : Hashtable
-        {
-            private readonly ComboBox _owningComboBoxBox;
-            private readonly ObjectIDGenerator _idGenerator = new ObjectIDGenerator();
-
-            public ComboBoxItemAccessibleObjectCollection(ComboBox owningComboBoxBox)
-            {
-                _owningComboBoxBox = owningComboBoxBox;
-            }
-
-            public override object this[object key]
-            {
-                get
-                {
-                    int id = GetId(key);
-                    if (!ContainsKey(id))
-                    {
-                        var itemAccessibleObject = new ComboBoxItemAccessibleObject(_owningComboBoxBox, key);
-                        base[id] = itemAccessibleObject;
-                    }
-
-                    return base[id];
-                }
-
-                set
-                {
-                    int id = GetId(key);
-                    base[id] = value;
-                }
-            }
-
-            public int GetId(object item)
-            {
-                return unchecked((int)_idGenerator.GetId(item, out var _));
-            }
-        }
-
-        /// <summary>
-        ///  ComboBox control accessible object with UI Automation provider functionality.
-        ///  This inherits from the base ComboBoxExAccessibleObject and ComboBoxAccessibleObject
-        ///  to have all base functionality.
-        /// </summary>
-        internal class ComboBoxAccessibleObject : ControlAccessibleObject
-        {
-            private const int COMBOBOX_ACC_ITEM_INDEX = 1;
-
-            private ComboBoxChildDropDownButtonUiaProvider _dropDownButtonUiaProvider;
-            private readonly ComboBox _owningComboBox;
-
-            /// <summary>
-            ///  Initializes new instance of ComboBoxAccessibleObject.
-            /// </summary>
-            /// <param name="owningComboBox">The owning ComboBox control.</param>
-            public ComboBoxAccessibleObject(ComboBox owningComboBox) : base(owningComboBox)
-            {
-                _owningComboBox = owningComboBox;
-                ItemAccessibleObjects = new ComboBoxItemAccessibleObjectCollection(owningComboBox);
-            }
-
-            private void ComboBoxDefaultAction(bool expand)
-            {
-                if (_owningComboBox.IsHandleCreated && _owningComboBox.DroppedDown != expand)
-                {
-                    _owningComboBox.DroppedDown = expand;
-                }
-            }
-
-            internal override bool IsIAccessibleExSupported()
-            {
-                if (_owningComboBox != null)
-                {
-                    return true;
-                }
-
-                return base.IsIAccessibleExSupported();
-            }
-
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.ExpandCollapsePatternId)
-                {
-                    if (_owningComboBox.DropDownStyle == ComboBoxStyle.Simple)
-                    {
-                        return false;
-                    }
-
-                    return true;
-                }
-
-                if (patternId == UiaCore.UIA.ValuePatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    if (_owningComboBox != null)
-                    {
-                        // we need to provide a unique ID
-                        // others are implementing this in the same manner
-                        // first item is static - 0x2a (RuntimeIDFirstItem)
-                        // second item can be anything, but here it is a hash
-
-                        var runtimeId = new int[3];
-                        runtimeId[0] = RuntimeIDFirstItem;
-                        runtimeId[1] = (int)(long)_owningComboBox.InternalHandle;
-                        runtimeId[2] = _owningComboBox.GetHashCode();
-
-                        return runtimeId;
-                    }
-
-                    return base.RuntimeId;
-                }
-            }
-
-            internal override void Expand()
-            {
-                ComboBoxDefaultAction(true);
-            }
-
-            internal override void Collapse()
-            {
-                ComboBoxDefaultAction(false);
-            }
-
-            internal override UiaCore.ExpandCollapseState ExpandCollapseState
-            {
-                get
-                {
-                    return _owningComboBox.IsHandleCreated && _owningComboBox.DroppedDown ? UiaCore.ExpandCollapseState.Expanded : UiaCore.ExpandCollapseState.Collapsed;
-                }
-            }
-
-            internal override string get_accNameInternal(object childID)
-            {
-                ValidateChildID(ref childID);
-
-                if (childID != null && ((int)childID) == COMBOBOX_ACC_ITEM_INDEX)
-                {
-                    return Name;
-                }
-
-                return base.get_accNameInternal(childID);
-            }
-
-            internal override string get_accKeyboardShortcutInternal(object childID)
-            {
-                ValidateChildID(ref childID);
-                if (childID != null && ((int)childID) == COMBOBOX_ACC_ITEM_INDEX)
-                {
-                    return KeyboardShortcut;
-                }
-
-                return base.get_accKeyboardShortcutInternal(childID);
-            }
-
-            /// <summary>
-            ///  Gets the collection of item accessible objects.
-            /// </summary>
-            public ComboBoxItemAccessibleObjectCollection ItemAccessibleObjects { get; }
-
-            /// <summary>
-            ///  Gets the DropDown button accessible object. (UI Automation provider)
-            /// </summary>
-            public ComboBoxChildDropDownButtonUiaProvider DropDownButtonUiaProvider
-            {
-                get
-                {
-                    if (_dropDownButtonUiaProvider is null)
-                    {
-                        _dropDownButtonUiaProvider = new ComboBoxChildDropDownButtonUiaProvider(_owningComboBox, _owningComboBox.InternalHandle);
-                    }
-
-                    return _dropDownButtonUiaProvider;
-                }
-            }
-
-            /// <summary>
-            ///  Returns the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                if (direction == UiaCore.NavigateDirection.FirstChild)
-                {
-                    return GetChildFragment(0);
-                }
-                else if (direction == UiaCore.NavigateDirection.LastChild)
-                {
-                    var childFragmentCount = GetChildFragmentCount();
-                    if (childFragmentCount > 0)
-                    {
-                        return GetChildFragment(childFragmentCount - 1);
-                    }
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            internal override UiaCore.IRawElementProviderSimple GetOverrideProviderForHwnd(IntPtr hwnd)
-            {
-                if (hwnd == _owningComboBox._childEdit.Handle)
-                {
-                    return _owningComboBox.ChildEditAccessibleObject;
-                }
-                else if (
-                    hwnd == _owningComboBox._childListBox.Handle ||
-                    hwnd == _owningComboBox._dropDownHandle)
-                {
-                    return _owningComboBox.ChildListAccessibleObject;
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Gets the accessible child corresponding to the specified index.
-            /// </summary>
-            /// <param name="index">The child index.</param>
-            /// <returns>The accessible child.</returns>
-            /// <remarks>
-            ///  GetChild method should be unchanged to not break the MSAA scenarios.
-            /// </remarks>
-            internal AccessibleObject GetChildFragment(int index)
-            {
-                if (_owningComboBox.DropDownStyle == ComboBoxStyle.DropDownList)
-                {
-                    if (index == 0)
-                    {
-                        return _owningComboBox.ChildTextAccessibleObject;
-                    }
-
-                    index--;
-                }
-
-                if (index == 0 && _owningComboBox.DropDownStyle != ComboBoxStyle.Simple)
-                {
-                    return DropDownButtonUiaProvider;
-                }
-
-                return null;
-            }
-
-            /// <summary>
-            ///  Gets the number of children belonging to an accessible object.
-            /// </summary>
-            /// <returns>The number of children.</returns>
-            /// <remarks>
-            ///  GetChildCount method should be unchanged to not break the MSAA scenarios.
-            /// </remarks>
-            internal int GetChildFragmentCount()
-            {
-                int childFragmentCount = 0;
-
-                if (_owningComboBox.DropDownStyle == ComboBoxStyle.DropDownList)
-                {
-                    childFragmentCount++; // Text instead of edit for style is DropDownList but not DropDown.
-                }
-
-                if (_owningComboBox.DropDownStyle != ComboBoxStyle.Simple)
-                {
-                    childFragmentCount++; // DropDown button.
-                }
-
-                return childFragmentCount;
-            }
-
-            /// <summary>
-            ///  Gets the accessible property value.
-            /// </summary>
-            /// <param name="propertyID">The accessible property ID.</param>
-            /// <returns>The accessible property value.</returns>
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.ComboBoxControlTypeId;
-                    case UiaCore.UIA.NamePropertyId:
-                        return Name;
-                    case UiaCore.UIA.HasKeyboardFocusPropertyId:
-                        return _owningComboBox.Focused;
-                    case UiaCore.UIA.NativeWindowHandlePropertyId:
-                        return _owningComboBox.InternalHandle;
-                    case UiaCore.UIA.IsExpandCollapsePatternAvailablePropertyId:
-                        return IsPatternSupported(UiaCore.UIA.ExpandCollapsePatternId);
-                    case UiaCore.UIA.IsValuePatternAvailablePropertyId:
-                        return IsPatternSupported(UiaCore.UIA.ValuePatternId);
-
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            internal void ResetListItemAccessibleObjects()
-            {
-                ItemAccessibleObjects.Clear();
-            }
-
-            internal void SetComboBoxItemFocus()
-            {
-                if (!_owningComboBox.IsHandleCreated)
-                {
-                    return;
-                }
-
-                var selectedItem = _owningComboBox.SelectedItem;
-                if (selectedItem is null)
-                {
-                    return;
-                }
-
-                if (ItemAccessibleObjects[selectedItem] is ComboBoxItemAccessibleObject itemAccessibleObject)
-                {
-                    itemAccessibleObject.SetFocus();
-                }
-            }
-
-            internal void SetComboBoxItemSelection()
-            {
-                if (!_owningComboBox.IsHandleCreated)
-                {
-                    return;
-                }
-
-                var selectedItem = _owningComboBox.SelectedItem;
-                if (selectedItem is null)
-                {
-                    return;
-                }
-
-                if (ItemAccessibleObjects[selectedItem] is ComboBoxItemAccessibleObject itemAccessibleObject)
-                {
-                    itemAccessibleObject.RaiseAutomationEvent(UiaCore.UIA.SelectionItem_ElementSelectedEventId);
-                }
-            }
-
-            internal override void SetFocus()
-            {
-                if (!_owningComboBox.IsHandleCreated)
-                {
-                    return;
-                }
-
-                base.SetFocus();
-
-                RaiseAutomationEvent(UiaCore.UIA.AutomationFocusChangedEventId);
-            }
-        }
-
-        /// <summary>
-        ///  Represents the ComboBox's child (inner) edit native window control accessible object with UI Automation provider functionality.
-        /// </summary>
-        internal class ComboBoxChildEditUiaProvider : ChildAccessibleObject
-        {
-            private const string COMBO_BOX_EDIT_AUTOMATION_ID = "1001";
-
-            private readonly ComboBox _owner;
-            private readonly IntPtr _handle;
-
-            /// <summary>
-            ///  Initializes new instance of ComboBoxChildEditUiaProvider.
-            /// </summary>
-            /// <param name="owner">The ComboBox owning control.</param>
-            /// <param name="childEditControlhandle">The child edit native window handle.</param>
-            public ComboBoxChildEditUiaProvider(ComboBox owner, IntPtr childEditControlhandle) : base(owner, childEditControlhandle)
-            {
-                _owner = owner;
-                _handle = childEditControlhandle;
-            }
-
-            /// <summary>
-            ///  Returns the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                switch (direction)
-                {
-                    case UiaCore.NavigateDirection.Parent:
-                        Debug.WriteLine("Edit parent " + _owner.AccessibilityObject.GetPropertyValue(UiaCore.UIA.ControlTypePropertyId));
-                        return _owner.AccessibilityObject;
-                    case UiaCore.NavigateDirection.NextSibling:
-                        if (_owner.DropDownStyle == ComboBoxStyle.Simple)
-                        {
-                            return null;
-                        }
-
-                        if (_owner.AccessibilityObject is ComboBoxAccessibleObject comboBoxAccessibleObject)
-                        {
-                            int comboBoxChildFragmentCount = comboBoxAccessibleObject.GetChildFragmentCount();
-                            if (comboBoxChildFragmentCount > 1)
-                            { // DropDown button is next;
-                                return comboBoxAccessibleObject.GetChildFragment(comboBoxChildFragmentCount - 1);
-                            }
-                        }
-
-                        return null;
-                    case UiaCore.NavigateDirection.PreviousSibling:
-                        comboBoxAccessibleObject = _owner.AccessibilityObject as ComboBoxAccessibleObject;
-                        if (comboBoxAccessibleObject != null)
-                        {
-                            AccessibleObject firstComboBoxChildFragment = comboBoxAccessibleObject.GetChildFragment(0);
-                            if (RuntimeId != firstComboBoxChildFragment.RuntimeId)
-                            {
-                                return firstComboBoxChildFragment;
-                            }
-                        }
-
-                        return null;
-                    default:
-                        return base.FragmentNavigate(direction);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the top level element.
-            /// </summary>
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owner.AccessibilityObject;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible property value.
-            /// </summary>
-            /// <param name="propertyID">The accessible property ID.</param>
-            /// <returns>The accessible property value.</returns>
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.RuntimeIdPropertyId:
-                        return RuntimeId;
-                    case UiaCore.UIA.BoundingRectanglePropertyId:
-                        return Bounds;
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.EditControlTypeId;
-                    case UiaCore.UIA.NamePropertyId:
-                        return Name ?? SR.ComboBoxEditDefaultAccessibleName;
-                    case UiaCore.UIA.AccessKeyPropertyId:
-                        return string.Empty;
-                    case UiaCore.UIA.HasKeyboardFocusPropertyId:
-                        return _owner.Focused;
-                    case UiaCore.UIA.IsKeyboardFocusablePropertyId:
-                        return (State & AccessibleStates.Focusable) == AccessibleStates.Focusable;
-                    case UiaCore.UIA.IsEnabledPropertyId:
-                        return _owner.Enabled;
-                    case UiaCore.UIA.AutomationIdPropertyId:
-                        return COMBO_BOX_EDIT_AUTOMATION_ID;
-                    case UiaCore.UIA.HelpTextPropertyId:
-                        return Help ?? string.Empty;
-                    case UiaCore.UIA.IsPasswordPropertyId:
-                        return false;
-                    case UiaCore.UIA.NativeWindowHandlePropertyId:
-                        return _handle;
-                    case UiaCore.UIA.IsOffscreenPropertyId:
-                        return false;
-
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            internal override UiaCore.IRawElementProviderSimple HostRawElementProvider
-            {
-                get
-                {
-                    UiaCore.UiaHostProviderFromHwnd(new HandleRef(this, _handle), out UiaCore.IRawElementProviderSimple provider);
-                    return provider;
-                }
-            }
-
-            internal override bool IsIAccessibleExSupported() => true;
-
-            /// <summary>
-            ///  Gets the runtime ID.
-            /// </summary>
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    var runtimeId = new int[2];
-                    runtimeId[0] = RuntimeIDFirstItem;
-                    runtimeId[1] = GetHashCode();
-
-                    return runtimeId;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  Represents the ComboBox's child (inner) list native window control accessible object with UI Automation provider functionality.
-        /// </summary>
-        internal class ComboBoxChildListUiaProvider : ChildAccessibleObject
-        {
-            private const string COMBO_BOX_LIST_AUTOMATION_ID = "1000";
-
-            private readonly ComboBox _owningComboBox;
-            private readonly IntPtr _childListControlhandle;
-
-            public ComboBoxChildListUiaProvider(ComboBox owningComboBox, IntPtr childListControlhandle) : base(owningComboBox, childListControlhandle)
-            {
-                _owningComboBox = owningComboBox;
-                _childListControlhandle = childListControlhandle;
-            }
-
-            /// <summary>
-            ///  Return the child object at the given screen coordinates.
-            /// </summary>
-            /// <param name="x">X coordinate.</param>
-            /// <param name="y">Y coordinate.</param>
-            /// <returns>The accessible object of corresponding element in the provided coordinates.</returns>
-            internal override UiaCore.IRawElementProviderFragment ElementProviderFromPoint(double x, double y)
-            {
-                var systemIAccessible = GetSystemIAccessibleInternal();
-                if (systemIAccessible != null)
-                {
-                    object result = systemIAccessible.accHitTest((int)x, (int)y);
-                    if (result is int childId)
-                    {
-                        return GetChildFragment(childId - 1);
-                    }
-                    else
-                    {
-                        return null;
-                    }
-                }
-
-                return base.ElementProviderFromPoint(x, y);
-            }
-
-            /// <summary>
-            ///  Request to return the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                switch (direction)
-                {
-                    case UiaCore.NavigateDirection.FirstChild:
-                        return GetChildFragment(0);
-                    case UiaCore.NavigateDirection.LastChild:
-                        var childFragmentCount = GetChildFragmentCount();
-                        if (childFragmentCount > 0)
-                        {
-                            return GetChildFragment(childFragmentCount - 1);
-                        }
-
-                        return null;
-                    default:
-                        return base.FragmentNavigate(direction);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the top level element.
-            /// </summary>
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owningComboBox.AccessibilityObject;
-                }
-            }
-
-            public AccessibleObject GetChildFragment(int index)
-            {
-                if (index < 0 || index >= _owningComboBox.Items.Count)
-                {
-                    return null;
-                }
-
-                var item = _owningComboBox.Items[index];
-                var comboBoxAccessibleObject = _owningComboBox.AccessibilityObject as ComboBoxAccessibleObject;
-                return comboBoxAccessibleObject.ItemAccessibleObjects[item] as AccessibleObject;
-            }
-
-            public int GetChildFragmentCount()
-            {
-                return _owningComboBox.Items.Count;
-            }
-
-            /// <summary>
-            ///  Gets the accessible property value.
-            /// </summary>
-            /// <param name="propertyID">The accessible property ID.</param>
-            /// <returns>The accessible property value.</returns>
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.RuntimeIdPropertyId:
-                        return RuntimeId;
-                    case UiaCore.UIA.BoundingRectanglePropertyId:
-                        return Bounds;
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.ListControlTypeId;
-                    case UiaCore.UIA.NamePropertyId:
-                        return Name;
-                    case UiaCore.UIA.AccessKeyPropertyId:
-                        return string.Empty;
-                    case UiaCore.UIA.HasKeyboardFocusPropertyId:
-                        return false; // Narrator should keep the keyboard focus on th ComboBox itself but not on the DropDown.
-                    case UiaCore.UIA.IsKeyboardFocusablePropertyId:
-                        return (State & AccessibleStates.Focusable) == AccessibleStates.Focusable;
-                    case UiaCore.UIA.IsEnabledPropertyId:
-                        return _owningComboBox.Enabled;
-                    case UiaCore.UIA.AutomationIdPropertyId:
-                        return COMBO_BOX_LIST_AUTOMATION_ID;
-                    case UiaCore.UIA.HelpTextPropertyId:
-                        return Help ?? string.Empty;
-                    case UiaCore.UIA.IsPasswordPropertyId:
-                        return false;
-                    case UiaCore.UIA.NativeWindowHandlePropertyId:
-                        return _childListControlhandle;
-                    case UiaCore.UIA.IsOffscreenPropertyId:
-                        return false;
-                    case UiaCore.UIA.IsSelectionPatternAvailablePropertyId:
-                        return true;
-                    case UiaCore.UIA.SelectionCanSelectMultiplePropertyId:
-                        return CanSelectMultiple;
-                    case UiaCore.UIA.SelectionIsSelectionRequiredPropertyId:
-                        return IsSelectionRequired;
-
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            internal override UiaCore.IRawElementProviderFragment GetFocus()
-            {
-                return GetFocused();
-            }
-
-            public override AccessibleObject GetFocused()
-            {
-                if (!_owningComboBox.IsHandleCreated)
-                {
-                    return null;
-                }
-
-                int selectedIndex = _owningComboBox.SelectedIndex;
-                return GetChildFragment(selectedIndex);
-            }
-
-            internal override UiaCore.IRawElementProviderSimple[] GetSelection()
-            {
-                if (!_owningComboBox.IsHandleCreated)
-                {
-                    return Array.Empty<UiaCore.IRawElementProviderSimple>();
-                }
-
-                int selectedIndex = _owningComboBox.SelectedIndex;
-
-                AccessibleObject itemAccessibleObject = GetChildFragment(selectedIndex);
-
-                if (itemAccessibleObject != null)
-                {
-                    return new UiaCore.IRawElementProviderSimple[] {
-                        itemAccessibleObject
-                    };
-                }
-
-                return Array.Empty<UiaCore.IRawElementProviderSimple>();
-            }
-
-            internal override bool CanSelectMultiple
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            internal override bool IsSelectionRequired
-            {
-                get
-                {
-                    return true;
-                }
-            }
-
-            /// <summary>
-            ///  Indicates whether specified pattern is supported.
-            /// </summary>
-            /// <param name="patternId">The pattern ID.</param>
-            /// <returns>True if specified </returns>
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.LegacyIAccessiblePatternId ||
-                    patternId == UiaCore.UIA.SelectionPatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            internal override UiaCore.IRawElementProviderSimple HostRawElementProvider
-            {
-                get
-                {
-                    UiaCore.UiaHostProviderFromHwnd(new HandleRef(this, _childListControlhandle), out UiaCore.IRawElementProviderSimple provider);
-                    return provider;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the runtime ID.
-            /// </summary>
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    var runtimeId = new int[3];
-                    runtimeId[0] = RuntimeIDFirstItem;
-                    runtimeId[1] = (int)(long)_owningComboBox.InternalHandle;
-                    runtimeId[2] = _owningComboBox.GetListNativeWindowRuntimeIdPart();
-
-                    return runtimeId;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible state.
-            /// </summary>
-            public override AccessibleStates State
-            {
-                get
-                {
-                    AccessibleStates state = AccessibleStates.Focusable;
-                    if (_owningComboBox.Focused)
-                    {
-                        state |= AccessibleStates.Focused;
-                    }
-
-                    return state;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  Represents the ComboBox's child text (is used instead of inner Edit when style is DropDownList but not DropDown) accessible object.
-        /// </summary>
-        internal class ComboBoxChildTextUiaProvider : AccessibleObject
-        {
-            private const int COMBOBOX_TEXT_ACC_ITEM_INDEX = 1;
-
-            private readonly ComboBox _owner;
-
-            /// <summary>
-            ///  Initializes new instance of ComboBoxChildTextUiaProvider.
-            /// </summary>
-            /// <param name="owner">The owning ComboBox control.</param>
-            public ComboBoxChildTextUiaProvider(ComboBox owner)
-            {
-                _owner = owner;
-            }
-
-            /// <summary>
-            ///  Gets the bounds.
-            /// </summary>
-            public override Rectangle Bounds
-            {
-                get
-                {
-                    return _owner.AccessibilityObject.Bounds;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the child ID.
-            /// </summary>
-            /// <returns>The child ID.</returns>
-            internal override int GetChildId()
-            {
-                return COMBOBOX_TEXT_ACC_ITEM_INDEX;
-            }
-
-            /// <summary>
-            ///  Gets or sets the accessible Name of ComboBox's child text element.
-            /// </summary>
-            public override string Name
-            {
-                get
-                {
-                    return _owner.AccessibilityObject.Name ?? string.Empty;
-                }
-                set
-                {
-                    // Do nothing.
-                }
-            }
-
-            /// <summary>
-            ///  Returns the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                switch (direction)
-                {
-                    case UiaCore.NavigateDirection.Parent:
-                        return _owner.AccessibilityObject;
-                    case UiaCore.NavigateDirection.NextSibling:
-                        if (_owner.AccessibilityObject is ComboBoxAccessibleObject comboBoxAccessibleObject)
-                        {
-                            int comboBoxChildFragmentCount = comboBoxAccessibleObject.GetChildFragmentCount();
-                            if (comboBoxChildFragmentCount > 1)
-                            { // DropDown button is next;
-                                return comboBoxAccessibleObject.GetChildFragment(comboBoxChildFragmentCount - 1);
-                            }
-                        }
-
-                        return null;
-                    case UiaCore.NavigateDirection.PreviousSibling:
-                        comboBoxAccessibleObject = _owner.AccessibilityObject as ComboBoxAccessibleObject;
-                        if (comboBoxAccessibleObject != null)
-                        {
-                            AccessibleObject firstComboBoxChildFragment = comboBoxAccessibleObject.GetChildFragment(0);
-                            if (RuntimeId != firstComboBoxChildFragment.RuntimeId)
-                            {
-                                return firstComboBoxChildFragment;
-                            }
-                        }
-
-                        return null;
-                    default:
-                        return base.FragmentNavigate(direction);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the top level element.
-            /// </summary>
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owner.AccessibilityObject;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible property value.
-            /// </summary>
-            /// <param name="propertyID">The accessible property ID.</param>
-            /// <returns>The accessible property value.</returns>
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.RuntimeIdPropertyId:
-                        return RuntimeId;
-                    case UiaCore.UIA.BoundingRectanglePropertyId:
-                        return Bounds;
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.TextControlTypeId;
-                    case UiaCore.UIA.NamePropertyId:
-                        return Name;
-                    case UiaCore.UIA.AccessKeyPropertyId:
-                        return string.Empty;
-                    case UiaCore.UIA.HasKeyboardFocusPropertyId:
-                        return _owner.Focused;
-                    case UiaCore.UIA.IsKeyboardFocusablePropertyId:
-                        return (State & AccessibleStates.Focusable) == AccessibleStates.Focusable;
-                    case UiaCore.UIA.IsEnabledPropertyId:
-                        return _owner.Enabled;
-                    case UiaCore.UIA.HelpTextPropertyId:
-                        return Help ?? string.Empty;
-                    case UiaCore.UIA.IsPasswordPropertyId:
-                    case UiaCore.UIA.IsOffscreenPropertyId:
-                        return false;
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the runtime ID.
-            /// </summary>
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    var runtimeId = new int[5];
-                    runtimeId[0] = RuntimeIDFirstItem;
-                    runtimeId[1] = (int)(long)_owner.InternalHandle;
-                    runtimeId[2] = _owner.GetHashCode();
-                    runtimeId[3] = GetHashCode();
-                    runtimeId[4] = GetChildId();
-
-                    return runtimeId;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible state.
-            /// </summary>
-            public override AccessibleStates State
-            {
-                get
-                {
-                    AccessibleStates state = AccessibleStates.Focusable;
-                    if (_owner.Focused)
-                    {
-                        state |= AccessibleStates.Focused;
-                    }
-
-                    return state;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  Represents the ComboBox child (inner) DropDown button accessible object with UI Automation functionality.
-        /// </summary>
-        internal class ComboBoxChildDropDownButtonUiaProvider : AccessibleObject
-        {
-            private const int COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX = 2;
-            private readonly ComboBox _owner;
-
-            /// <summary>
-            ///  Initializes new instance of ComboBoxChildDropDownButtonUiaProvider.
-            /// </summary>
-            /// <param name="owner">The owning ComboBox control.</param>
-            /// <param name="comboBoxControlhandle">The owning ComboBox control's handle.</param>
-            public ComboBoxChildDropDownButtonUiaProvider(ComboBox owner, IntPtr comboBoxControlhandle)
-            {
-                _owner = owner;
-                UseStdAccessibleObjects(comboBoxControlhandle);
-            }
-
-            /// <summary>
-            ///  Gets or sets the accessible Name of ComboBox's child DropDown button. ("Open" or "Close" depending on stat of the DropDown)
-            /// </summary>
-            public override string Name
-            {
-                get
-                {
-                    return get_accNameInternal(COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX);
-                }
-                set
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    systemIAccessible.set_accName(COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX, value);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the DropDown button bounds.
-            /// </summary>
-            public override Rectangle Bounds
-            {
-                get
-                {
-                    int left = 0;
-                    int top = 0;
-                    int width = 0;
-                    int height = 0;
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    systemIAccessible.accLocation(out left, out top, out width, out height, COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX);
-                    return new Rectangle(left, top, width, height);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the DropDown button default action.
-            /// </summary>
-            public override string DefaultAction
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    return systemIAccessible.accDefaultAction[COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX];
-                }
-            }
-
-            /// <summary>
-            ///  Returns the element in the specified direction.
-            /// </summary>
-            /// <param name="direction">Indicates the direction in which to navigate.</param>
-            /// <returns>Returns the element in the specified direction.</returns>
-            internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-            {
-                if (direction == UiaCore.NavigateDirection.Parent)
-                {
-                    return _owner.AccessibilityObject;
-                }
-                else if (direction == UiaCore.NavigateDirection.PreviousSibling)
-                {
-                    if (_owner.AccessibilityObject is ComboBoxAccessibleObject comboBoxAccessibleObject)
-                    {
-                        int comboBoxChildFragmentCount = comboBoxAccessibleObject.GetChildFragmentCount();
-                        if (comboBoxChildFragmentCount > 1)
-                        { // Text or edit is previous;
-                            return comboBoxAccessibleObject.GetChildFragment(comboBoxChildFragmentCount - 1);
-                        }
-                    }
-
-                    return null;
-                }
-
-                return base.FragmentNavigate(direction);
-            }
-
-            /// <summary>
-            ///  Gets the top level element.
-            /// </summary>
-            internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-            {
-                get
-                {
-                    return _owner.AccessibilityObject;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the child accessible object ID.
-            /// </summary>
-            /// <returns>The child accessible object ID.</returns>
-            internal override int GetChildId()
-            {
-                return COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX;
-            }
-
-            /// <summary>
-            ///  Gets the accessible property value.
-            /// </summary>
-            /// <param name="propertyID">The accessible property ID.</param>
-            /// <returns>The accessible property value.</returns>
-            internal override object GetPropertyValue(UiaCore.UIA propertyID)
-            {
-                switch (propertyID)
-                {
-                    case UiaCore.UIA.RuntimeIdPropertyId:
-                        return RuntimeId;
-                    case UiaCore.UIA.BoundingRectanglePropertyId:
-                        return BoundingRectangle;
-                    case UiaCore.UIA.ControlTypePropertyId:
-                        return UiaCore.UIA.ButtonControlTypeId;
-                    case UiaCore.UIA.NamePropertyId:
-                        return Name;
-                    case UiaCore.UIA.AccessKeyPropertyId:
-                        return KeyboardShortcut;
-                    case UiaCore.UIA.HasKeyboardFocusPropertyId:
-                        return _owner.Focused;
-                    case UiaCore.UIA.IsKeyboardFocusablePropertyId:
-                        return (State & AccessibleStates.Focusable) == AccessibleStates.Focusable;
-                    case UiaCore.UIA.IsEnabledPropertyId:
-                        return _owner.Enabled;
-                    case UiaCore.UIA.HelpTextPropertyId:
-                        return Help ?? string.Empty;
-                    case UiaCore.UIA.IsPasswordPropertyId:
-                        return false;
-                    case UiaCore.UIA.IsOffscreenPropertyId:
-                        return (State & AccessibleStates.Offscreen) == AccessibleStates.Offscreen;
-                    default:
-                        return base.GetPropertyValue(propertyID);
-                }
-            }
-
-            /// <summary>
-            ///  Gets the help text.
-            /// </summary>
-            public override string Help
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    return systemIAccessible.accHelp[COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX];
-                }
-            }
-
-            /// <summary>
-            ///  Gets the keyboard shortcut.
-            /// </summary>
-            public override string KeyboardShortcut
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    return systemIAccessible.get_accKeyboardShortcut(COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX);
-                }
-            }
-
-            /// <summary>
-            ///  Indicates whether specified pattern is supported.
-            /// </summary>
-            /// <param name="patternId">The pattern ID.</param>
-            /// <returns>True if specified </returns>
-            internal override bool IsPatternSupported(UiaCore.UIA patternId)
-            {
-                if (patternId == UiaCore.UIA.LegacyIAccessiblePatternId ||
-                    patternId == UiaCore.UIA.InvokePatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            /// <summary>
-            ///  Gets the accessible role.
-            /// </summary>
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    var accRole = systemIAccessible.get_accRole(COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX);
-                    return accRole != null
-                        ? (AccessibleRole)accRole
-                        : AccessibleRole.None;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the runtime ID.
-            /// </summary>
-            internal override int[] RuntimeId
-            {
-                get
-                {
-                    var runtimeId = new int[5];
-                    runtimeId[0] = RuntimeIDFirstItem;
-                    runtimeId[1] = (int)(long)_owner.InternalHandle;
-                    runtimeId[2] = _owner.GetHashCode();
-
-                    // Made up constant from MSAA proxy. When MSAA proxy is used as an accessibility provider,
-                    // the similar Runtime ID is returned (for consistency purpose)
-                    const int generatedRuntimeId = 61453;
-                    runtimeId[3] = generatedRuntimeId;
-                    runtimeId[4] = COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX;
-                    return runtimeId;
-                }
-            }
-
-            /// <summary>
-            ///  Gets the accessible state.
-            /// </summary>
-            public override AccessibleStates State
-            {
-                get
-                {
-                    var systemIAccessible = GetSystemIAccessibleInternal();
-                    var accState = systemIAccessible.get_accState(COMBOBOX_DROPDOWN_BUTTON_ACC_ITEM_INDEX);
-                    return accState != null
-                        ? (AccessibleStates)accState
-                        : AccessibleStates.None;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  This subclasses an autocomplete window so that we can determine if control is inside the AC wndproc.
-        /// </summary>
-        private sealed class ACNativeWindow : NativeWindow
-        {
-            internal static int s_inWndProcCnt;
-
-            // This hashtable can contain null for those ACWindows we find, but are sure are not ours.
-            private static readonly Hashtable s_ACWindows = new Hashtable();
-
-            internal ACNativeWindow(IntPtr acHandle)
-            {
-                Debug.Assert(!s_ACWindows.ContainsKey(acHandle));
-                AssignHandle(acHandle);
-                s_ACWindows.Add(acHandle, this);
-                EnumChildWindows(new HandleRef(this, acHandle),
-                    ACNativeWindow.RegisterACWindowRecursive);
-            }
-
-            private static BOOL RegisterACWindowRecursive(IntPtr handle)
-            {
-                if (!s_ACWindows.ContainsKey(handle))
-                {
-                    ACNativeWindow newAC = new ACNativeWindow(handle);
-                }
-                return BOOL.TRUE;
-            }
-
-            internal bool Visible => IsWindowVisible(this).IsTrue();
-
-            static internal bool AutoCompleteActive
-            {
-                get
-                {
-                    if (s_inWndProcCnt > 0)
-                    {
-                        return true;
-                    }
-                    foreach (object o in s_ACWindows.Values)
-                    {
-                        if (o is ACNativeWindow window && window.Visible)
-                        {
-                            return true;
-                        }
-                    }
-                    return false;
-                }
-            }
-
-            protected override void WndProc(ref Message m)
-            {
-                s_inWndProcCnt++;
-                try
-                {
-                    base.WndProc(ref m);
-                }
-                finally
-                {
-                    s_inWndProcCnt--;
-                }
-
-                if (m.Msg == (int)WM.NCDESTROY)
-                {
-                    Debug.Assert(s_ACWindows.ContainsKey(Handle));
-                    s_ACWindows.Remove(Handle);   //so we do not leak ac windows.
-                }
-            }
-
-            internal static void RegisterACWindow(IntPtr acHandle, bool subclass)
-            {
-                if (subclass && s_ACWindows.ContainsKey(acHandle))
-                {
-                    if (s_ACWindows[acHandle] is null)
-                    {
-                        s_ACWindows.Remove(acHandle); //if an external handle got destroyed, dont let it stop us.
-                    }
-                }
-
-                if (!s_ACWindows.ContainsKey(acHandle))
-                {
-                    if (subclass)
-                    {
-                        ACNativeWindow newAC = new ACNativeWindow(acHandle);
-                    }
-                    else
-                    {
-                        s_ACWindows.Add(acHandle, null);
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  This method clears out null entries so we get a clean BEFORE and AFTER snapshot
-            ///  null entries are ACWindows that belong to someone else.
-            /// </summary>
-            internal static void ClearNullACWindows()
-            {
-                ArrayList nulllist = new ArrayList();
-                foreach (DictionaryEntry e in s_ACWindows)
-                {
-                    if (e.Value is null)
-                    {
-                        nulllist.Add(e.Key);
-                    }
-                }
-                foreach (IntPtr handle in nulllist)
-                {
-                    s_ACWindows.Remove(handle);
-                }
-            }
-        }
-
-        /// <summary>
-        ///  This finds all autcomplete windows that belong to the active thread.
-        /// </summary>
-        private class AutoCompleteDropDownFinder
-        {
-            private const int MaxClassName = 256;
-            private const string AutoCompleteClassName = "Auto-Suggest Dropdown";
-            private bool _shouldSubClass;
-
-            internal void FindDropDowns()
-            {
-                FindDropDowns(true);
-            }
-
-            internal void FindDropDowns(bool subclass)
-            {
-                if (!subclass)
-                {
-                    //generating a before snapshot -- lets lose the null handles
-                    ACNativeWindow.ClearNullACWindows();
-                }
-
-                // Look for a popped up dropdown
-                _shouldSubClass = subclass;
-                EnumThreadWindows(
-                    Kernel32.GetCurrentThreadId(),
-                    Callback);
-            }
-
-            private BOOL Callback(IntPtr hWnd)
-            {
-                HandleRef hRef = new HandleRef(null, hWnd);
-
-                // Check class name and see if it's visible
-                if (GetClassName(hRef) == AutoCompleteClassName)
-                {
-                    ACNativeWindow.RegisterACWindow(hRef.Handle, _shouldSubClass);
-                }
-
-                return BOOL.TRUE;
-            }
-
-            static string GetClassName(HandleRef hRef)
-            {
-                StringBuilder sb = new StringBuilder(MaxClassName);
-                UnsafeNativeMethods.GetClassName(hRef, sb, MaxClassName);
-                return sb.ToString();
-            }
-        }
-
         private FlatComboAdapter FlatComboBoxAdapter
         {
             get
@@ -6164,223 +3923,6 @@ namespace System.Windows.Forms
 
         internal virtual FlatComboAdapter CreateFlatComboAdapterInstance()
             => new FlatComboAdapter(this, smallButton: false);
-
-        internal class FlatComboAdapter
-        {
-            private Rectangle _outerBorder;
-            private Rectangle _innerBorder;
-            private Rectangle _innerInnerBorder;
-            internal Rectangle _dropDownRect;
-            private Rectangle _whiteFillRect;
-            private Rectangle _clientRect;
-            private readonly RightToLeft _origRightToLeft; // The combo box's RTL value when we were created
-
-            private const int WhiteFillRectWidth = 5; // used for making the button look smaller than it is
-
-            private const int OFFSET_2PIXELS = 2;
-            protected static int s_offsetPixels = OFFSET_2PIXELS;
-
-            public FlatComboAdapter(ComboBox comboBox, bool smallButton)
-            {
-                // adapter is re-created when combobox is resized, see IsValid method, thus we don't need to handle DPI changed explicitly
-                s_offsetPixels = comboBox.LogicalToDeviceUnits(OFFSET_2PIXELS);
-
-                _clientRect = comboBox.ClientRectangle;
-                int dropDownButtonWidth = SystemInformation.GetHorizontalScrollBarArrowWidthForDpi(comboBox._deviceDpi);
-                _outerBorder = new Rectangle(_clientRect.Location, new Size(_clientRect.Width - 1, _clientRect.Height - 1));
-                _innerBorder = new Rectangle(_outerBorder.X + 1, _outerBorder.Y + 1, _outerBorder.Width - dropDownButtonWidth - 2, _outerBorder.Height - 2);
-                _innerInnerBorder = new Rectangle(_innerBorder.X + 1, _innerBorder.Y + 1, _innerBorder.Width - 2, _innerBorder.Height - 2);
-                _dropDownRect = new Rectangle(_innerBorder.Right + 1, _innerBorder.Y, dropDownButtonWidth, _innerBorder.Height + 1);
-
-                // fill in several pixels of the dropdown rect with white so that it looks like the combo button is thinner.
-                if (smallButton)
-                {
-                    _whiteFillRect = _dropDownRect;
-                    _whiteFillRect.Width = WhiteFillRectWidth;
-                    _dropDownRect.X += WhiteFillRectWidth;
-                    _dropDownRect.Width -= WhiteFillRectWidth;
-                }
-
-                _origRightToLeft = comboBox.RightToLeft;
-
-                if (_origRightToLeft == RightToLeft.Yes)
-                {
-                    _innerBorder.X = _clientRect.Width - _innerBorder.Right;
-                    _innerInnerBorder.X = _clientRect.Width - _innerInnerBorder.Right;
-                    _dropDownRect.X = _clientRect.Width - _dropDownRect.Right;
-                    _whiteFillRect.X = _clientRect.Width - _whiteFillRect.Right + 1;  // since we're filling, we need to move over to the next px.
-                }
-            }
-
-            public bool IsValid(ComboBox combo)
-            {
-                return (combo.ClientRectangle == _clientRect && combo.RightToLeft == _origRightToLeft);
-            }
-
-            /// <summary>
-            ///  Paints over the edges of the combo box to make it appear flat.
-            /// </summary>
-            public virtual void DrawFlatCombo(ComboBox comboBox, Graphics g)
-            {
-                if (comboBox.DropDownStyle == ComboBoxStyle.Simple)
-                {
-                    return;
-                }
-
-                Color outerBorderColor = GetOuterBorderColor(comboBox);
-                Color innerBorderColor = GetInnerBorderColor(comboBox);
-                bool rightToLeft = comboBox.RightToLeft == RightToLeft.Yes;
-
-                // Draw the drop down
-                DrawFlatComboDropDown(comboBox, g, _dropDownRect);
-
-                // When we are disabled there is one line of color that seems to eek through if backcolor is set
-                // so lets erase it.
-                if (!LayoutUtils.IsZeroWidthOrHeight(_whiteFillRect))
-                {
-                    // Fill in two more pixels with white so it looks smaller.
-                    using var b = innerBorderColor.GetCachedSolidBrushScope();
-                    g.FillRectangle(b, _whiteFillRect);
-                }
-
-                // Draw the outer border
-
-                using var outerBorderPen = outerBorderColor.GetCachedPenScope();
-                g.DrawRectangle(outerBorderPen, _outerBorder);
-                if (rightToLeft)
-                {
-                    g.DrawRectangle(
-                        outerBorderPen,
-                        new Rectangle(_outerBorder.X, _outerBorder.Y, _dropDownRect.Width + 1, _outerBorder.Height));
-                }
-                else
-                {
-                    g.DrawRectangle(
-                        outerBorderPen,
-                        new Rectangle(_dropDownRect.X, _outerBorder.Y, _outerBorder.Right - _dropDownRect.X, _outerBorder.Height));
-                }
-
-                // Draw the inner border
-                using var innerBorderPen = innerBorderColor.GetCachedPenScope();
-                g.DrawRectangle(innerBorderPen, _innerBorder);
-                g.DrawRectangle(innerBorderPen, _innerInnerBorder);
-
-                // Draw a dark border around everything if we're in popup mode
-                if ((!comboBox.Enabled) || (comboBox.FlatStyle == FlatStyle.Popup))
-                {
-                    bool focused = comboBox.ContainsFocus || comboBox.MouseIsOver;
-                    Color borderPenColor = GetPopupOuterBorderColor(comboBox, focused);
-
-                    using var borderPen = borderPenColor.GetCachedPenScope();
-                    Pen innerPen = comboBox.Enabled ? borderPen : SystemPens.Control;
-
-                    // Around the dropdown
-                    if (rightToLeft)
-                    {
-                        g.DrawRectangle(
-                            innerPen,
-                            new Rectangle(_outerBorder.X, _outerBorder.Y, _dropDownRect.Width + 1, _outerBorder.Height));
-                    }
-                    else
-                    {
-                        g.DrawRectangle(
-                            innerPen,
-                            new Rectangle(_dropDownRect.X, _outerBorder.Y, _outerBorder.Right - _dropDownRect.X, _outerBorder.Height));
-                    }
-
-                    // Around the whole combobox.
-                    g.DrawRectangle(borderPen, _outerBorder);
-                }
-            }
-
-            /// <summary>
-            ///  Paints over the edges of the combo box to make it appear flat.
-            /// </summary>
-            protected virtual void DrawFlatComboDropDown(ComboBox comboBox, Graphics g, Rectangle dropDownRect)
-            {
-                g.FillRectangle(SystemBrushes.Control, dropDownRect);
-
-                Brush brush = (comboBox.Enabled) ? SystemBrushes.ControlText : SystemBrushes.ControlDark;
-
-                Point middle = new Point(dropDownRect.Left + dropDownRect.Width / 2, dropDownRect.Top + dropDownRect.Height / 2);
-                if (_origRightToLeft == RightToLeft.Yes)
-                {
-                    // if the width is odd - favor pushing it over one pixel left.
-                    middle.X -= (dropDownRect.Width % 2);
-                }
-                else
-                {
-                    // if the width is odd - favor pushing it over one pixel right.
-                    middle.X += (dropDownRect.Width % 2);
-                }
-
-                g.FillPolygon(
-                    brush,
-                    new Point[]
-                    {
-                        new Point(middle.X - s_offsetPixels, middle.Y - 1),
-                        new Point(middle.X + s_offsetPixels + 1, middle.Y - 1),
-                        new Point(middle.X, middle.Y + s_offsetPixels)
-                    });
-            }
-
-            protected virtual Color GetOuterBorderColor(ComboBox comboBox)
-                => comboBox.Enabled ? SystemColors.Window : SystemColors.ControlDark;
-
-            protected virtual Color GetPopupOuterBorderColor(ComboBox comboBox, bool focused)
-            {
-                if (!comboBox.Enabled)
-                {
-                    return SystemColors.ControlDark;
-                }
-                return focused ? SystemColors.ControlDark : SystemColors.Window;
-            }
-
-            protected virtual Color GetInnerBorderColor(ComboBox comboBox)
-                => comboBox.Enabled ? comboBox.BackColor : SystemColors.Control;
-
-            /// <summary>
-            ///  This eliminates flicker by removing the pieces we're going to paint ourselves from the update region.
-            ///  Note the UpdateRegionBox is the bounding box of the actual update region. This is here so we can
-            ///  quickly eliminate rectangles that arent in the update region.
-            /// </summary>
-            public unsafe void ValidateOwnerDrawRegions(ComboBox comboBox, Rectangle updateRegionBox)
-            {
-                if (comboBox != null)
-                {
-                    return;
-                }
-
-                Rectangle topOwnerDrawArea = new Rectangle(0, 0, comboBox.Width, _innerBorder.Top);
-                Rectangle bottomOwnerDrawArea = new Rectangle(0, _innerBorder.Bottom, comboBox.Width, comboBox.Height - _innerBorder.Bottom);
-                Rectangle leftOwnerDrawArea = new Rectangle(0, 0, _innerBorder.Left, comboBox.Height);
-                Rectangle rightOwnerDrawArea = new Rectangle(_innerBorder.Right, 0, comboBox.Width - _innerBorder.Right, comboBox.Height);
-
-                if (topOwnerDrawArea.IntersectsWith(updateRegionBox))
-                {
-                    RECT validRect = new RECT(topOwnerDrawArea);
-                    ValidateRect(comboBox, &validRect);
-                }
-
-                if (bottomOwnerDrawArea.IntersectsWith(updateRegionBox))
-                {
-                    RECT validRect = new RECT(bottomOwnerDrawArea);
-                    ValidateRect(comboBox, &validRect);
-                }
-
-                if (leftOwnerDrawArea.IntersectsWith(updateRegionBox))
-                {
-                    RECT validRect = new RECT(leftOwnerDrawArea);
-                    ValidateRect(comboBox, &validRect);
-                }
-
-                if (rightOwnerDrawArea.IntersectsWith(updateRegionBox))
-                {
-                    RECT validRect = new RECT(rightOwnerDrawArea);
-                    ValidateRect(comboBox, &validRect);
-                }
-            }
-        }
 
         /// <summary>
         ///  Represents the ComboBox child native window type.
